fix: address CI failures in batched-fix PR
- InitiativeList.test.tsx: update 4 assertions on the heading text
  to match the new translated value "Lista de Iniciativas".
- Split party colors out of parties.ts into a new parties-colors.ts
  module so tests can import them without triggering the eager
  supabase env-var check in parties.ts. parties.ts re-exports
  PARTY_COLORS and getPartyColor for backwards compatibility.
